+++ test/cases/resolve-assets-split-chunk/src/assets/js/module-b.js
@@ -1,0 +1,6 @@
+const lib = require('./lib');
+const ansis = require('ansis');
+let value = lib.methodB();
+value = ansis.strip(value);
+
+module.exports = value;